+++ src/screens/simulation/PublicSimulation.js
@@ -1,10 +1,9 @@
 import React, { useState, useEffect, useRef, useCallback } from 'react';
 import { Form, Field } from 'react-final-form';
-import { makeStyles, Card, Typography, Slider, useMediaQuery, useTheme } from '@material-ui/core';
+import { makeStyles, Card, Typography, useMediaQuery, useTheme } from '@material-ui/core';
 import AwesomeDebouncePromise from 'awesome-debounce-promise';
 import { debounce } from 'lodash';
 import { format } from 'date-fns';
-import { ZoomIn, ZoomOut } from '@material-ui/icons';
 
 import { formatParametersForModel, defaultParameters, extractGraphTimeframes } from './common';
 import api from '../../api';
@@ -19,6 +18,7 @@ import ProportionField from '../../components/fields/ProportionField';
 import AutoSave from '../../components/fields/AutoSave';
 import PublicDescriptionModal from './PublicDescriptionModal';
 import colors from './colors';
+import { ZoomSlider } from './ZoomSlider';
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -114,24 +114,6 @@ const useStyles = makeStyles((theme) => ({
             alignItems: 'center',
             minWidth: 200,
             marginBottom: 12,
-        },
-    },
-    zoom: {
-        display: 'flex',
-        flexDirection: 'column',
-        height: '100%',
-        [theme.breakpoints.down('sm')]: {
-            flexDirection: 'row',
-        },
-        alignItems: 'center',
-        justifyContent: 'space-between',
-    },
-    zoomSlider: {
-        [theme.breakpoints.down('sm')]: {
-            width: '85% !important',
-        },
-        [theme.breakpoints.up('sm')]: {
-            height: '80% !important',
         },
     },
 }));
@@ -330,42 +312,6 @@ const getTimeframesFromValues = ({
         enabled: true,
     },
 ];
-
-const ZoomSlider = ({ onChange, initValue, min, max }) => {
-    var logMin = Math.log(min);
-    var logMax = Math.log(max);
-    const scale = (logMax - logMin) / 100;
-    const zoomToValue = (value) => Math.exp(logMin + scale * value);
-    const valueToZoom = (value) => (Math.log(value) - logMin) / scale;
-
-    const [innerMax, setInnerMax] = useState(valueToZoom(initValue));
-    const classes = useStyles();
-
-    const theme = useTheme();
-    const small = useMediaQuery(theme.breakpoints.down('sm'));
-
-    const handleChange = (commited) => (_, value) => {
-        setInnerMax(value);
-        commited && onChange(zoomToValue(value));
-    };
-
-    return (
-        <div className={classes.zoom}>
-            <ZoomOut fontSize="small" />
-            <Slider
-                classes={{ root: classes.zoomSlider, vertical: classes.zoomSlider }}
-                orientation={small ? 'horizontal' : 'vertical'}
-                value={innerMax}
-                max={100}
-                min={0}
-                onChangeCommitted={handleChange(true)}
-                onChange={handleChange(false)}
-                aria-labelledby="range-slider"
-            />
-            <ZoomIn fontSize="small" />
-        </div>
-    );
-};
 
 const PublicSimulation = () => {
     const classes = useStyles();
